requirements.txt
fastapi
uvicorn
python-dotenv
chromadb
google-generativeai
PyPDF2
pydantic
multipart